[WIP] Leitura de valor int16 signed com multiplicador corrigida; erro na leitura da corrente

+++ cliente_modbus/clientemodbus.py
@@ -122,7 +122,7 @@ class ClienteMODBUS():
                 else: # int16
                     raw_val = resp.registers[0]
                     signed_val = raw_val if raw_val < 32768 else raw_val - 65536
-                    return signed_val/multiplicador
+                    return int(signed_val/multiplicador)
                 
             case TipoEndereco.COIL:
                 return self._cliente.read_coils(addr,1)[0]
